Create dispatch handshake customer side
The protocol here is similar to a handshake. When the customer requests
a dispatch the firebase will be updated with a dispatch request. Then
the customer will talk to the closest driver and initiate a handshake.
If the closest driver accepts it will handshake back and then the two
devices will connect, if the driver declines or is timed out the
customer will move on to the next closest driver.

The driver side of the handshake still needs to be written.